+++ mirtop/libs/parse.py
@@ -70,7 +70,7 @@ def add_subparser_gff(subparsers):
     parser.add_argument("--hairpin", help="hairpin.fa")
     parser.add_argument("--gtf", help="gtf/gff file with precursor and mature position to genome.")
     parser.add_argument("--format", help="Input format, default BAM file.",
-                        choices=['BAM', 'seqbuster', 'srnabench', 'prost', 'isomirsea'], default="BAM")
+                        choices=['BAM', 'seqbuster', 'srnabench', 'prost', 'isomirsea', 'gff'], default="BAM")
     parser.add_argument("--out-format", help="Supported formats: gff3 or gtf", default = "gtf")
     parser.add_argument("--add-extra", help="Add extra attributes to gff", action="store_true")
     parser = _add_debug_option(parser)